Fixed _apistatus view for 0.9.11 rabbitmq/redis status
Sensu appears to have changed the `/info` api, to the extent that the response no longer contains a 'health' object. Given the _apistatus.html.haml view looks for this, a web request to /api/status generates a 500. I've put this small change in for 0.9.11, but this will probably break for 0.9.10 now.